Deprecated bdist_wheel and updated the README
