+++ README.rst
@@ -1,19 +1,24 @@
 wheel
 =====
 
-This library is the reference implementation of the Python wheel packaging
-standard, as defined in `PEP 427`_.
+This is a command line tool for manipulating Python wheel files, as defined in
+`PEP 427`_. It contains the following functionality:
 
-It has two different roles:
-
-#. A setuptools_ extension for building wheels that provides the
-   ``bdist_wheel`` setuptools command
-#. A command line tool for working with wheel files
-
-It should be noted that wheel is **not** intended to be used as a library, and
-as such there is no stable, public API.
+* Convert ``.egg`` archives into ``.whl``
+* Unpack wheel archives
+* Repack wheel archives
+* Add or remove tags in existing wheel archives
 
 .. _PEP 427: https://www.python.org/dev/peps/pep-0427/
+
+Historical note
+---------------
+
+This project used to contain the implementation of the setuptools_ ``bdist_wheel``
+command, but as of setuptools v70.1, it no longer needs ``wheel`` installed for that to
+work. Thus, you should install this **only** if you intend to use the ``wheel`` command
+line tool!
+
 .. _setuptools: https://pypi.org/project/setuptools/
 
 Documentation
