+++ src/wheel/cli/convert.py
@@ -7,7 +7,7 @@ import tempfile
 import zipfile
 from glob import iglob
 
-from ..bdist_wheel import bdist_wheel
+from .._bdist_wheel import bdist_wheel
 from ..wheelfile import WheelFile
 from . import WheelError
