+++ tests/test_macosx_libfile.py
@@ -5,7 +5,7 @@ import struct
 import sysconfig
 
 import pytest
-from wheel.bdist_wheel import get_platform
+from wheel._bdist_wheel import get_platform
 from wheel.macosx_libfile import extract_macosx_min_system_version
